requirements.txt
tensorflow
streamlit
matplotlib
seaborn
Pillow
scikit-learn
